refactor: remove json model namespace
