+++ Bucketeer2/Sources/Internal/Evaluation/EvaluationDao.swift
@@ -1,7 +1,5 @@
 import Foundation
 
-typealias Evaluation = JSON.Evaluation
-
 protocol EvaluationDao {
     func put(userId: String, evaluations: [Evaluation]) throws
     func get(userId: String) throws -> [Evaluation]
