--- Bucketeer2/Sources/Internal/JSON/JSON.Duration.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension JSON {
-    struct Duration: Codable, Hashable {
-        let seconds: Int64
-    }
-}
