--- Bucketeer2/Sources/Internal/JSON/JSON.Reason.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-extension JSON {
-    struct Reason: Codable, Hashable {
-        let type: ReasonType
-        var rule_id: String = ""
-    }
-}
