--- Bucketeer2/Sources/Internal/JSON/JSON.User.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension JSON {
-    struct User: Codable, Hashable {
-        let id: String
-        var data: [String: String]
-        // note: tagged_data is not used in client SDK
-    }
-}
