--- Bucketeer2/Sources/Internal/JSON/JSON.UserData.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension JSON {
-    struct UserData {
-        let value: [String: String]
-    }
-}
