--- Bucketeer2/Sources/Internal/JSON/JSON.UserEvaluations.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-extension JSON {
-    struct UserEvaluations: Codable {
-        var id: String
-        var evaluations: [Evaluation]
-    }
-}
