--- Bucketeer2/Sources/Internal/JSON/JSON.Variation.swift
@@ -1,10 +1,0 @@
-import Foundation
-
-extension JSON {
-    struct Variation: Codable, Hashable {
-        let id: String
-        var value: String
-        let name: String?
-        let description: String?
-    }
-}
