--- Bucketeer2/Sources/Internal/JSON/JSON.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-/// namespace
-struct JSON {
-    private init() {}
-}
