+++ Bucketeer2/Sources/Internal/Model/Duration.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct Duration: Codable, Hashable {
+    let seconds: Int64
+}
