+++ Bucketeer2/Sources/Internal/Model/Reason.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Reason: Codable, Hashable {
+    let type: ReasonType
+    var rule_id: String = ""
+}
