+++ Bucketeer2/Sources/Internal/Model/UserData.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct UserData {
+    let value: [String: String]
+}
