+++ Bucketeer2/Sources/Internal/Model/UserEvaluations.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct UserEvaluations: Codable {
+    var id: String
+    var evaluations: [Evaluation]
+}
