+++ Bucketeer2/Sources/Internal/Model/Variation.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct Variation: Codable, Hashable {
+    let id: String
+    var value: String
+    let name: String?
+    let description: String?
+}
