+++ Bucketeer2/Sources/Internal/Remote/ApiClient.swift
@@ -1,12 +1,12 @@
 import Foundation
 
 protocol ApiClient {
-    func getEvaluations(user: JSON.User, userEvaluationsId: String, timeoutMillis: Int64?, completion: ((GetEvaluationsResult) -> Void)?)
+    func getEvaluations(user: User, userEvaluationsId: String, timeoutMillis: Int64?, completion: ((GetEvaluationsResult) -> Void)?)
     func registerEvents(events: [Event], completion: ((Result<RegisterEventsResponse, BKTError>) -> Void)?)
 }
 
 extension ApiClient {
-    func getEvaluations(user: JSON.User, userEvaluationsId: String, completion: ((GetEvaluationsResult) -> Void)?) {
+    func getEvaluations(user: User, userEvaluationsId: String, completion: ((GetEvaluationsResult) -> Void)?) {
         getEvaluations(
             user: user,
             userEvaluationsId: userEvaluationsId,
