+++ Bucketeer2/Sources/Internal/Remote/RequestBody/GetEvaluationsRequestBody.swift
@@ -2,7 +2,7 @@ import Foundation
 
 struct GetEvaluationsRequestBody: Codable {
     let tag: String
-    let user: JSON.User
+    let user: User
     let user_evaluations_id: String
-    let source_id: JSON.SourceID
+    let source_id: SourceID
 }
