+++ Bucketeer2Tests/EventInteractorTests.swift
@@ -27,7 +27,7 @@ final class EventInteractorTests: XCTestCase {
                     timestamp: 1,
                     feature_id: Evaluation.mock1.feature_id,
                     feature_version: Evaluation.mock1.feature_version,
-                    user_id: JSON.User.mock1.id,
+                    user_id: User.mock1.id,
                     variation_id: Evaluation.mock1.variation_id,
                     user: .mock1,
                     reason: Evaluation.mock1.reason,
@@ -72,7 +72,7 @@ final class EventInteractorTests: XCTestCase {
                 event: .evaluation(.init(
                     timestamp: 1,
                     feature_id: "featureId1",
-                    user_id: JSON.User.mock1.id,
+                    user_id: User.mock1.id,
                     user: .mock1,
                     reason: .init(type: .client),
                     tag: "featureTag1",
@@ -116,7 +116,7 @@ final class EventInteractorTests: XCTestCase {
                 event: .goal(.init(
                     timestamp: 1,
                     goal_id: "goalId1",
-                    user_id: JSON.User.mock1.id,
+                    user_id: User.mock1.id,
                     value: 1,
                     user: .mock1,
                     tag: "featureTag1",
