+++ Bucketeer2Tests/Mock/MockEvents.swift
@@ -7,7 +7,7 @@ extension Event {
         event: .goal(.init(
             timestamp: 1,
             goal_id: "goal1",
-            user_id: JSON.User.mock1.id,
+            user_id: User.mock1.id,
             value: 1,
             user: .mock1,
             tag: "tag1",
@@ -21,7 +21,7 @@ extension Event {
         event: .goal(.init(
             timestamp: 1,
             goal_id: "goal2",
-            user_id: JSON.User.mock1.id,
+            user_id: User.mock1.id,
             value: 1,
             user: .mock1,
             tag: "tag2",
@@ -36,7 +36,7 @@ extension Event {
             timestamp: 1,
             feature_id: "feature1",
             feature_version: 1,
-            user_id: JSON.User.mock1.id,
+            user_id: User.mock1.id,
             variation_id: "variation1",
             user: .mock1,
             reason: .init(type: .rule, rule_id: "rule1"),
@@ -52,7 +52,7 @@ extension Event {
             timestamp: 1,
             feature_id: "feature2",
             feature_version: 1,
-            user_id: JSON.User.mock1.id,
+            user_id: User.mock1.id,
             variation_id: "variation2",
             user: .mock1,
             reason: .init(type: .rule, rule_id: "rule2"),
