+++ Bucketeer2Tests/Mock/MockUsers.swift
@@ -1,7 +1,7 @@
 import Foundation
 @testable import Bucketeer2
 
-extension JSON.User {
-    static let mock1 = JSON.User(id: "user1", data: ["age": "28"])
-    static let mock2 = JSON.User(id: "user2", data: [:])
+extension User {
+    static let mock1 = User(id: "user1", data: ["age": "28"])
+    static let mock2 = User(id: "user2", data: [:])
 }
